GroupDetail: delete container, add Firestore
